Reorganize / sort layout and size modifiers
These CSS properties have been growing a bit organically
(read: disorderly), so took a moment to clean them up by
sorting them.
