+++ frontend/kobweb-compose/src/jsMain/kotlin/com/varabyte/kobweb/compose/ui/modifiers/LayoutModifiers.kt
@@ -18,6 +18,10 @@ fun Modifier.aspectRatio(ratio: AspectRatio): Modifier = styleModifier {
     aspectRatio(ratio)
 }
 
+fun Modifier.clear(clear: Clear) = styleModifier {
+    clear(clear)
+}
+
 fun Modifier.lineHeight(value: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
     lineHeight(value)
 }
@@ -28,99 +32,6 @@ fun Modifier.lineHeight(value: Number): Modifier = styleModifier {
 
 fun Modifier.lineHeight(lineHeight: LineHeight): Modifier = styleModifier {
     lineHeight(lineHeight)
-}
-
-fun Modifier.fillMaxWidth(percent: CSSLengthOrPercentageNumericValue = 100.percent) = styleModifier {
-    width(percent)
-}
-
-fun Modifier.fillMaxHeight(percent: CSSPercentageNumericValue = 100.percent) = styleModifier {
-    height(percent)
-}
-
-fun Modifier.fillMaxSize(percent: CSSPercentageNumericValue = 100.percent): Modifier = styleModifier {
-    width(percent)
-    height(percent)
-}
-
-fun Modifier.size(size: CSSLengthOrPercentageNumericValue): Modifier = size(width = size, height = size)
-
-fun Modifier.size(width: CSSLengthOrPercentageNumericValue, height: CSSLengthOrPercentageNumericValue): Modifier =
-    styleModifier {
-    width(width)
-    height(height)
-}
-
-fun Modifier.minSize(size: CSSLengthOrPercentageNumericValue): Modifier = minSize(width = size, height = size)
-
-fun Modifier.minSize(width: CSSLengthOrPercentageNumericValue, height: CSSLengthOrPercentageNumericValue): Modifier =
-    styleModifier {
-    minWidth(width)
-    minHeight(height)
-}
-
-fun Modifier.maxSize(size: CSSLengthOrPercentageNumericValue): Modifier = maxSize(width = size, height = size)
-
-fun Modifier.maxSize(width: CSSLengthOrPercentageNumericValue, height: CSSLengthOrPercentageNumericValue): Modifier =
-    styleModifier {
-    maxWidth(width)
-    maxHeight(height)
-}
-
-fun Modifier.width(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    width(size)
-}
-
-fun Modifier.height(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    height(size)
-}
-
-fun Modifier.width(width: Width): Modifier = styleModifier {
-    width(width)
-}
-
-fun Modifier.width(auto: CSSAutoKeyword): Modifier = styleModifier {
-    width(auto)
-}
-
-fun Modifier.height(height: Height): Modifier = styleModifier {
-    height(height)
-}
-
-fun Modifier.height(auto: CSSAutoKeyword): Modifier = styleModifier {
-    height(auto)
-}
-
-fun Modifier.minWidth(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    minWidth(size)
-}
-
-fun Modifier.minWidth(minWidth: MinWidth): Modifier = styleModifier {
-    minWidth(minWidth)
-}
-
-fun Modifier.maxWidth(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    maxWidth(size)
-}
-
-fun Modifier.maxWidth(maxWidth: MaxWidth): Modifier = styleModifier {
-    maxWidth(maxWidth)
-}
-
-fun Modifier.minHeight(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    minHeight(size)
-}
-
-fun Modifier.minHeight(minHeight: MinHeight): Modifier = styleModifier {
-    minHeight(minHeight)
-}
-
-fun Modifier.maxHeight(size: CSSLengthOrPercentageNumericValue): Modifier = styleModifier {
-    maxHeight(size)
-}
-
-fun Modifier.maxHeight(maxHeight: MaxHeight): Modifier = styleModifier {
-    maxHeight(maxHeight)
 }
 
 class MarginScope internal constructor(private val styleScope: StyleScope) {
@@ -176,8 +87,32 @@ fun Modifier.marginBlock(
     marginBlock(start, end)
 }
 
-fun Modifier.resize(resize: Resize) = styleModifier {
-    resize(resize)
+fun Modifier.overflow(overflow: Overflow) = styleModifier {
+    overflow(overflow)
+}
+
+fun Modifier.overflow(overflowX: Overflow, overflowY: Overflow) = styleModifier {
+    overflow(overflowX, overflowY)
+}
+
+class OverflowScope internal constructor(private val styleScope: StyleScope) {
+    fun x(overflowX: Overflow) = styleScope.overflowX(overflowX)
+    fun y(overflowY: Overflow) = styleScope.overflowY(overflowY)
+}
+
+fun Modifier.overflow(scope: OverflowScope.() -> Unit) = styleModifier {
+    OverflowScope(this).scope()
+}
+
+// TODO(#168): Remove these before v1.0
+@Deprecated("Use overflow { x(overflowX) } instead.", ReplaceWith("overflow { x(overflowX) }"))
+fun Modifier.overflowX(overflowX: Overflow) = overflow { x(overflowX) }
+
+@Deprecated("Use overflow { y(overflowY) } instead.", ReplaceWith("overflow { y(overflowY) }"))
+fun Modifier.overflowY(overflowY: Overflow) = overflow { y(overflowY) }
+
+fun Modifier.overflowWrap(overflowWrap: OverflowWrap) = styleModifier {
+    overflowWrap(overflowWrap)
 }
 
 class PaddingScope internal constructor(private val styleScope: StyleScope) {
@@ -265,32 +200,8 @@ fun Modifier.paddingBlockStart(value: CSSLengthOrPercentageNumericValue) = paddi
 @Deprecated("Use paddingBlock { end(value) } instead.", ReplaceWith("paddingBlock { end(value) }"))
 fun Modifier.paddingBlockEnd(value: CSSLengthOrPercentageNumericValue) = paddingBlock { end(value) }
 
-fun Modifier.overflow(overflow: Overflow) = styleModifier {
-    overflow(overflow)
-}
-
-fun Modifier.overflow(overflowX: Overflow, overflowY: Overflow) = styleModifier {
-    overflow(overflowX, overflowY)
-}
-
-class OverflowScope internal constructor(private val styleScope: StyleScope) {
-    fun x(overflowX: Overflow) = styleScope.overflowX(overflowX)
-    fun y(overflowY: Overflow) = styleScope.overflowY(overflowY)
-}
-
-fun Modifier.overflow(scope: OverflowScope.() -> Unit) = styleModifier {
-    OverflowScope(this).scope()
-}
-
-// TODO(#168): Remove these before v1.0
-@Deprecated("Use overflow { x(overflowX) } instead.", ReplaceWith("overflow { x(overflowX) }"))
-fun Modifier.overflowX(overflowX: Overflow) = overflow { x(overflowX) }
-
-@Deprecated("Use overflow { y(overflowY) } instead.", ReplaceWith("overflow { y(overflowY) }"))
-fun Modifier.overflowY(overflowY: Overflow) = overflow { y(overflowY) }
-
-fun Modifier.overflowWrap(overflowWrap: OverflowWrap) = styleModifier {
-    overflowWrap(overflowWrap)
+fun Modifier.resize(resize: Resize) = styleModifier {
+    resize(resize)
 }
 
 fun Modifier.verticalAlign(verticalAlign: VerticalAlign) = styleModifier {
@@ -303,8 +214,4 @@ fun Modifier.verticalAlign(value: CSSLengthOrPercentageNumericValue) = styleModi
 
 fun Modifier.zIndex(value: Number) = styleModifier {
     zIndex(value)
-}
-
-fun Modifier.clear(clear: Clear) = styleModifier {
-    clear(clear)
 }
